requirements.txt
Pillow==6.2.0
PyQt5==5.13.1
PyQt5-sip==4.19.19
